refactor(PieceResponse): Domain -> View Dto Mapper 적용
