+++ src/main/java/view/dto/PieceResponse.java
@@ -1,0 +1,8 @@
+package view.dto;
+
+public record PieceResponse(
+        int file,
+        int rank,
+        char role
+) {
+}
